src/cloud_functions/ripp-scraper/requirements.txt
functions-framework==3.*
requests~=2.31.0
google-cloud-storage==2.13.0
google-cloud-bigquery==3.13.0